refactor: リソースファイルを ui に集めた
