+++ core/ui/src/main/java/app/kaito_dogi/mybrary/core/ui/component/BookImage.kt
@@ -36,8 +36,8 @@ fun BookImage(
     model = imageUrl.value,
     contentDescription = "${title}の表紙",
     modifier = modifier.aspectRatio(bookAspectRatio),
-    placeholder = painterResource(id = R.drawable.ui_book_image_placeholder),
-    error = painterResource(id = R.drawable.ui_book_image_placeholder),
+    placeholder = painterResource(id = R.drawable.img_book_placeholder),
+    error = painterResource(id = R.drawable.img_book_placeholder),
     onLoading = onLoading,
     onSuccess = onSuccess,
     onError = onError,
