+++ core/ui/src/main/res/drawable-night/img_book_placeholder.xml
