+++ core/ui/src/main/res/drawable/icon_add.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M11,13H5V11H11V5H13V11H19V13H13V19H11V13Z"
+        android:fillColor="#ffffff"/>
+  </group>
+</vector>
