+++ core/ui/src/main/res/drawable/icon_archive.xml
