+++ core/ui/src/main/res/drawable/icon_arrow_back.xml
