+++ core/ui/src/main/res/drawable/icon_heart_outlined.xml
