+++ core/ui/src/main/res/drawable/icon_send.xml
