+++ core/ui/src/main/res/drawable/img_book_placeholder.xml
